Add files via upload
Hi, 
I add a new folder for the code of the plotter because I don't know if it fits with the output section so to not disorganize the repository I create a new folder, I also would like to add a comment on the read me section so the user know the purpose of this plotter but I still working on how to do that. 
Hope I did it right and this upload the folder.